refactor loss
