+++ src/numpitron/nn/__init__.py
@@ -1,7 +1,7 @@
 # flake8: noqa
 from numpitron.nn.activation import ReLU, Softmax, softmax
 from numpitron.nn.attention import Attention
-from numpitron.nn.cross_entropy import SoftmaxCrossEntropy
+from numpitron.nn.loss import SoftmaxCrossEntropy
 from numpitron.nn.embedding import InputEmbedding, OutputEmbedding, PositionalEmbedding
 from numpitron.nn.linear import Linear
 from numpitron.nn.mlp import MLP
